Fixed PHP builtin webserver lacks of 404 error for not '.php' files

+++ jail/default_scripts/php_run.sh
@@ -9,14 +9,36 @@
 check_program php5
 check_program x-www-browser
 if [ -f index.php ] ; then
+    #Configure session
 	SESSIONPATH=$HOME/.php_sessions
 	mkdir $SESSIONPATH
 	cp /etc/php5/cli/php.ini .php.ini
-	cat >> .php.ini <<FIN
+	#Generate php.ini
+	cat >> .php.ini <<END_OF_INI
 	
-session.save_path ="$SESSIONPATH"
-
-FIN
+session.save_path="$SESSIONPATH"
+error_reporting=E_ALL
+display_errors=On
+display_startup_errors=On
+END_OF_INI
+    #Generate router
+    cat >> .router.php << 'END_OF_PHP'
+<?php $path='.'.urldecode(parse_url($_SERVER["REQUEST_URI"],PHP_URL_PATH));
+if(is_file($path) || is_file($path.'/index.php') || is_file($path.'/index.html') ){
+    unset($path);
+    return false;
+}
+$pclean=htmlentities($path);
+http_response_code(404);
+header(':', true, 404);
+?>
+<!doctype html>
+<html><head><title>404 Not found</title>
+<style>h1{background-color: aqua;text-align:center} code{font-size:150%}</style>
+</head>
+<body><h1>404 Not found</h1><p>The requested resource <code><?php echo "'$pclean'"; ?></code> 
+was not found on this server</body></html>
+END_OF_PHP
     cat > vpl_wexecution <<"END_OF_SCRIPT"
 #!/bin/bash
 while true; do
@@ -24,7 +46,7 @@ while true; do
    	lsof -i :$PHPPORT
    	[ "$?" != "0" ] && break
 done
-php -c .php.ini -d display_errors=On -S "127.0.0.1:$PHPPORT" &
+php -c .php.ini -S "127.0.0.1:$PHPPORT" .router.php &
 x-www-browser "127.0.0.1:$PHPPORT"
 END_OF_SCRIPT
     chmod +x vpl_wexecution
